# infra/cloud/openapi/v1.snapshot.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nexora.Api.Cloud",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/public/branding": {
      "get": {
        "tags": [
          "Branding"
        ],
        "parameters": [
          {
            "name": "host",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/branding.webmanifest": {
      "get": {
        "tags": [
          "Branding"
        ],
        "parameters": [
          {
            "name": "host",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": { }
              },
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/branding": {
      "get": {
        "tags": [
          "Branding"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Branding"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBrandingResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBrandingResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBrandingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenant/branding/logo": {
      "post": {
        "tags": [
          "Branding"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadBrandingAssetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadBrandingAssetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadBrandingAssetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBrandingAssetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBrandingAssetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBrandingAssetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrandingColorsDto": {
        "type": "object",
        "properties": {
          "primary": {
            "type": "string",
            "nullable": true
          },
          "secondary": {
            "type": "string",
            "nullable": true
          },
          "surface": {
            "type": "string",
            "nullable": true
          },
          "onPrimary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingColorsPatch": {
        "type": "object",
        "properties": {
          "primary": {
            "type": "string",
            "nullable": true
          },
          "secondary": {
            "type": "string",
            "nullable": true
          },
          "surface": {
            "type": "string",
            "nullable": true
          },
          "onPrimary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingContrastIssueResponse": {
        "type": "object",
        "properties": {
          "pair": {
            "type": "string",
            "nullable": true
          },
          "ratio": {
            "type": "number",
            "format": "double"
          },
          "suggested": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingContrastResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "minimumRatio": {
            "type": "number",
            "format": "double"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandingContrastIssueResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingDto": {
        "type": "object",
        "properties": {
          "colors": {
            "$ref": "#/components/schemas/BrandingColorsDto"
          },
          "logo": {
            "$ref": "#/components/schemas/BrandingLogoDto"
          },
          "favicon": {
            "type": "string",
            "nullable": true
          },
          "fonts": {
            "$ref": "#/components/schemas/BrandingFontsDto"
          },
          "radius": {
            "type": "integer",
            "format": "int32"
          },
          "texts": {
            "$ref": "#/components/schemas/BrandingTextsDto"
          },
          "pwa": {
            "$ref": "#/components/schemas/BrandingPwaDto"
          }
        },
        "additionalProperties": false
      },
      "BrandingFontsDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "nullable": true
          },
          "display": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingFontsPatch": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "nullable": true
          },
          "display": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingLogoDto": {
        "type": "object",
        "properties": {
          "light": {
            "type": "string",
            "nullable": true
          },
          "dark": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingPwaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "themeColor": {
            "type": "string",
            "nullable": true
          },
          "icons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandingPwaIconDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingPwaIconDto": {
        "type": "object",
        "properties": {
          "src": {
            "type": "string",
            "nullable": true
          },
          "sizes": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "purpose": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingPwaPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "themeColor": {
            "type": "string",
            "nullable": true
          },
          "icons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandingPwaIconDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingResponse": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantBrandingInfoResponse"
          },
          "branding": {
            "$ref": "#/components/schemas/BrandingDto"
          },
          "configVersion": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BrandingTextsDto": {
        "type": "object",
        "properties": {
          "welcome": {
            "type": "string",
            "nullable": true
          },
          "orderConfirmed": {
            "type": "string",
            "nullable": true
          },
          "thanks": {
            "type": "string",
            "nullable": true
          },
          "terms": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BrandingTextsPatch": {
        "type": "object",
        "properties": {
          "welcome": {
            "type": "string",
            "nullable": true
          },
          "orderConfirmed": {
            "type": "string",
            "nullable": true
          },
          "thanks": {
            "type": "string",
            "nullable": true
          },
          "terms": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorResponse"
            },
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "recoverable": {
            "type": "boolean"
          },
          "requiresAuthorization": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "FieldErrorResponse": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TenantBrandingInfoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBrandingRequest": {
        "type": "object",
        "properties": {
          "colors": {
            "$ref": "#/components/schemas/BrandingColorsPatch"
          },
          "logo": {
            "$ref": "#/components/schemas/BrandingLogoDto"
          },
          "favicon": {
            "type": "string",
            "nullable": true
          },
          "fonts": {
            "$ref": "#/components/schemas/BrandingFontsPatch"
          },
          "radius": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "texts": {
            "$ref": "#/components/schemas/BrandingTextsPatch"
          },
          "pwa": {
            "$ref": "#/components/schemas/BrandingPwaPatch"
          }
        },
        "additionalProperties": false
      },
      "UpdateBrandingResponse": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantBrandingInfoResponse"
          },
          "branding": {
            "$ref": "#/components/schemas/BrandingDto"
          },
          "configVersion": {
            "type": "integer",
            "format": "int32"
          },
          "contrast": {
            "$ref": "#/components/schemas/BrandingContrastResponse"
          }
        },
        "additionalProperties": false
      },
...
      "UploadBrandingAssetRequest": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "bytes": {
            "type": "integer",
            "format": "int32"
          },
          "sha256": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UploadBrandingAssetResponse": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadUrl": {
            "type": "string",
            "nullable": true
          },
          "publicUrl": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}